Убрал валидатор из базового CRUD

+++ app/crud/base.py
@@ -7,6 +7,10 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from app.models import User
 
 
+from http import HTTPStatus
+from fastapi import HTTPException
+
+
 class CRUDBase:
     """ Класс-основа для операций CRUD. """
 
@@ -75,6 +79,11 @@ class CRUDBase:
             session: AsyncSession,
     ):
         """ Базовый метод удаления объекта. """
+        if db_obj.fully_invested or db_obj.invested_amount > 0:
+            raise HTTPException(
+                status_code=HTTPStatus.BAD_REQUEST,
+                detail='В проект были внесены средства, не подлежит удалению!'
+            )
         await session.delete(db_obj)
         await session.commit()
         return db_obj
